requirements.txt
Django==3.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
Pillow==10.0.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0 